+++ packages/js/src/RoomSessionDevice.ts
